rename files and update docs
